+++ back/src/index.js
@@ -1,0 +1,28 @@
+import express from "express";
+import { graphqlHTTP } from "express-graphql";
+//import  schema  from "./graphql/schema"; 
+import { dbConnection } from "./database/config";
+
+const app = express();
+//require('dotenv').config();
+dbConnection();
+
+
+// app.use(express.static('public'));
+// app.use(express.json());
+
+app.use('/api/usuarios', require('./routes/usuarios'));
+app.use('/api/proyectos', require('./routes/proyectos'));
+app.use('/api/inscripciones', require('./routes/inscripciones'));
+app.use('/api/avances', require('./routes/avances'));
+
+const schema = {};
+app.use('/graphql', graphqlHTTP({
+    graphiql: true,
+    schema: schema
+}))
+
+
+app.listen(4000, () => {
+    console.log('Servidor corriendo en el puerto 4000')
+});
